+++ frontend/src/pages/detail/hooks/useHighlightData.ts
@@ -1,5 +1,4 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-import { useEffect } from 'react';
 import { HighlightType } from '../types/highlight';
 import {
   deleteHighlight,
@@ -8,7 +7,7 @@ import {
   postHighlight,
 } from '@/apis/highlight';
 
-export const useHighlightManager = ({ articleId }: { articleId: number }) => {
+export const useHighlightData = ({ articleId }: { articleId: number }) => {
   const queryClient = useQueryClient();
   const { data: highlights } = useQuery({
     queryKey: ['highlight'],
@@ -57,36 +56,6 @@ export const useHighlightManager = ({ articleId }: { articleId: number }) => {
   const updateMemo = (id: number, memo: string) => {
     updateHighlight({ id, data: { memo } });
   };
-
-  useEffect(() => {
-    const handleMouseOver = (e: Event) => {
-      const target = e.target as HTMLElement;
-      if (target.tagName === 'MARK' && target.dataset.highlightId) {
-        const id = target.dataset.highlightId;
-        document
-          .querySelectorAll(`mark[data-highlight-id="${id}"]`)
-          .forEach((el) => el.classList.add('hovered-highlight'));
-      }
-    };
-
-    const handleMouseOut = (e: Event) => {
-      const target = e.target as HTMLElement;
-      if (target.tagName === 'MARK' && target.dataset.highlightId) {
-        const id = target.dataset.highlightId;
-        document
-          .querySelectorAll(`mark[data-highlight-id="${id}"]`)
-          .forEach((el) => el.classList.remove('hovered-highlight'));
-      }
-    };
-
-    document.addEventListener('mouseover', handleMouseOver);
-    document.addEventListener('mouseout', handleMouseOut);
-
-    return () => {
-      document.removeEventListener('mouseover', handleMouseOver);
-      document.removeEventListener('mouseout', handleMouseOut);
-    };
-  }, []);
 
   return {
     highlights,
