+++ frontend/src/pages/detail/hooks/useHighlightHoverEffect.ts
@@ -1,0 +1,32 @@
+import { useEffect } from 'react';
+
+export const useHighlightHoverEffect = () => {
+  useEffect(() => {
+    const handleMouseOver = (e: Event) => {
+      const target = e.target as HTMLElement;
+      if (target.tagName === 'MARK' && target.dataset.highlightId) {
+        const id = target.dataset.highlightId;
+        document
+          .querySelectorAll(`mark[data-highlight-id="${id}"]`)
+          .forEach((el) => el.classList.add('hovered-highlight'));
+      }
+    };
+
+    const handleMouseOut = (e: Event) => {
+      const target = e.target as HTMLElement;
+      if (target.tagName === 'MARK' && target.dataset.highlightId) {
+        const id = target.dataset.highlightId;
+        document
+          .querySelectorAll(`mark[data-highlight-id="${id}"]`)
+          .forEach((el) => el.classList.remove('hovered-highlight'));
+      }
+    };
+
+    document.addEventListener('mouseover', handleMouseOver);
+    document.addEventListener('mouseout', handleMouseOut);
+    return () => {
+      document.removeEventListener('mouseover', handleMouseOver);
+      document.removeEventListener('mouseout', handleMouseOut);
+    };
+  }, []);
+};
